+++ src/cli.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-async-promise-executor */
 import fs from 'fs';
 import { program } from 'commander';
 import forge from 'node-forge';
@@ -12,9 +13,9 @@ const pkg: { version: string } = JSON.parse(fs.readFileSync(path.join(__dirname,
 
 program.version(pkg.version);
 
-async function getClientDevice(serial?: string): Promise<DeviceClient> {
+async function getClientDevice(serials: string[]): Promise<DeviceClient[]> {
   const adb = Adb.createClient();
-  if (!serial) {
+  if (!serials || !serials.length) {
     const devices: Device[] = await adb.listDevices();
     if (devices.length == 0) {
       console.error('no Android devices found');
@@ -22,14 +23,17 @@ async function getClientDevice(serial?: string): Promise<DeviceClient> {
       return;
     }
     if (devices.length == 1) {
-      return devices[0].getClient();
+      return [devices[0].getClient()];
     }
     console.error('Multiple devices avaliable provide a serial number to choose one');
     console.error(devices.map(d => `- ${d.id}`).join('\n'));
     process.exit(1);
-    return;
   }
-  return adb.getDevice(serial)
+  if (serials.length === 1 && (serials[0] === 'all' || serials[0] === '*')) {
+    serials = (await adb.listDevices()).map((d: Device) => d.id);
+  }
+
+  return serials.map(serial => adb.getDevice(serial))
 }
 
 program
@@ -82,99 +86,117 @@ program
 
 
 program
-  .command('usb-tethering-on [serial]')
+  .command('usb-tethering-on [...serials]')
   .description('Enable USB tethering.')
-  .action(async (serial?: string) => {
-    const device = await getClientDevice(serial);
-    if (await device.extra.usbTethering(true)) {
-      console.log('tethering enabled');
-      await device.extra.back();
-    } else {
-      console.log('failed or already enabled');
+  .action(async (...serials: string[]) => {
+    const devices = await getClientDevice(serials);
+    const process = async (device: DeviceClient) => {
+      if (await device.extra.usbTethering(true)) {
+        console.log(`[${device.serial}] tethering enabled`);
+        await device.extra.back();
+      } else {
+        console.log(`[${device.serial}] failed or already enabled`);
+      }
     }
+    await Promise.all(devices.map(process));
   });
 
 program
-  .command('usb-tethering-off [serial]')
-  .description('Enable USB tethering.')
-  .action(async (serial?: string) => {
-    const device = await getClientDevice(serial);
-    if (await device.extra.usbTethering(false)) {
-      console.log('tethering enabled');
-      await device.extra.back();
-    } else {
-      console.log('failed or already enabled');
+  .command('usb-tethering-off [...serials]')
+  .description('Disable USB tethering.')
+  .action(async (...serials: string[]) => {
+    const devices = await getClientDevice(serials);
+    const process = async (device: DeviceClient) => {
+      if (await device.extra.usbTethering(false)) {
+        console.log(`[${device.serial}] tethering enabled`);
+        await device.extra.back();
+      } else {
+        console.log(`[${device.serial}] failed or already enabled`);
+      }
     }
+    await Promise.all(devices.map(process));
   });
 
 program
-  .command('airplane-mode-on [serial]')
+  .command('airplane-mode-on [...serials]')
   .description('Enable airplane.')
-  .action(async (serial?: string) => {
-    const device = await getClientDevice(serial);
-    if (await device.extra.airPlainMode(true)) {
-      console.log('airplane enabled');
-      await device.extra.back();
-    } else {
-      console.log('airplane or already enabled');
-      process.exit(1);
+  .action(async (...serials: string[]) => {
+    const devices = await getClientDevice(serials);
+    const process = async (device: DeviceClient) => {
+      if (await device.extra.airPlainMode(true)) {
+        console.log(`[${device.serial}] airplane enabled`);
+        await device.extra.back();
+      } else {
+        console.log(`[${device.serial}] airplane or already enabled`);
+      }
     }
+    await Promise.all(devices.map(process));
   });
 
 program
-  .command('airplane-mode-off [serial]')
+  .command('airplane-mode-off [...serials]')
   .description('Disable airplane.')
-  .action(async (serial?: string) => {
-    const device = await getClientDevice(serial);
-    if (await device.extra.airPlainMode(false)) {
-      console.log('airplane disabled');
-      await device.extra.back();
-    } else {
-      console.log('airplane or already enabled');
-      process.exit(1);
+  .action(async (...serials: string[]) => {
+    const devices = await getClientDevice(serials);
+    const process = async (device: DeviceClient) => {
+      if (await device.extra.airPlainMode(false)) {
+        console.log(`[${device.serial}] airplane disabled`);
+        await device.extra.back();
+      } else {
+        console.log(`[${device.serial}] airplane or already enabled`);
+      }
     }
+    await Promise.all(devices.map(process));
   });
 
 program
-  .command('airplane-mode-on-off [serial]')
+  .command('airplane-mode-on-off [...serials]')
   .description('Enable then Disable airplane.')
-  .action(async (serial?: string) => {
-    const device = await getClientDevice(serial);
-    await device.extra.airPlainMode(true)
-    await device.extra.back();
-    if (await device.extra.airPlainMode(false)) {
-      console.log('airplane disabled');
+  .action(async (...serials: string[]) => {
+    const devices = await getClientDevice(serials);
+    const process = async (device: DeviceClient) => {
+      await device.extra.airPlainMode(true)
       await device.extra.back();
-    } else {
-      console.log('airplane or already enabled');
-      process.exit(1);
+      if (await device.extra.airPlainMode(false)) {
+        console.log(`[${device.serial}] airplane disabled`);
+        await device.extra.back();
+      } else {
+        console.log(`[${device.serial}] airplane or already enabled`);
+      }
     }
+    await Promise.all(devices.map(process));
   });
 
 
 for (const type of ['bluetooth', 'data', 'wifi'] as const) {
   program
-    .command(`${type}-off [serial]`)
+    .command(`${type}-off [...serials]`)
     .description('Disable bluetooth.')
-    .action(async (serial?: string) => {
-      const device = await getClientDevice(serial);
-      if (await device.extra.setSvc(type, false)) {
-        console.log(`${type} disabled`);
-      } else {
-        console.log(`failed or ${type} already disabled`);
+    .action(async (...serials: string[]) => {
+      const devices = await getClientDevice(serials);
+      const process = async (device: DeviceClient) => {
+        if (await device.extra.setSvc(type, false)) {
+          console.log(`[${device.serial}] ${type} disabled`);
+        } else {
+          console.log(`[${device.serial}] failed ${type} already disabled`);
+        }
       }
+      await Promise.all(devices.map(process));
     });
 
   program
-    .command(`${type}-on [serial]`)
+    .command(`${type}-on [...serials]`)
     .description('Enable bluetooth.')
-    .action(async (serial?: string) => {
-      const device = await getClientDevice(serial);
-      if (await device.extra.setSvc(type, true)) {
-        console.log('Bluetooth enabled');
-      } else {
-        console.log('failed or already enabled');
+    .action(async (...serials: string[]) => {
+      const devices = await getClientDevice(serials);
+      const process = async (device: DeviceClient) => {
+        if (await device.extra.setSvc(type, true)) {
+          console.log(`[${device.serial}] ${type} enabled`);
+        } else {
+          console.log(`[${device.serial}] failed ${type} already enabled`);
+        }
       }
+      await Promise.all(devices.map(process));
     });
 }