+++ tasks/realTest.ts
@@ -384,6 +384,8 @@ const main = async () => {
   // console.log(ret1);
   // const ret2 = await deviceClient.stat2('/');
   // console.log(ret2);
+  // const ret2 = await deviceClient.readdir2('/');
+  // console.log(ret2);
   // await deviceClient.extra.usbTethering(true);
   // await deviceClient.extra.airPlainMode(true);
   // await deviceClient.extra.airPlainMode(false);